restructure hasura tests
