+++ Dockerfile-dev
@@ -1,8 +1,8 @@
 FROM python:3.8-slim AS hasura-build
 
-COPY ./backend/database-service/tests/requirements.txt .
+COPY ./backend/requirements.txt .
 
-RUN apt update && apt install -y wget \
+RUN apt update && apt install -y gcc wget \
   && wget https://raw.githubusercontent.com/vishnubob/wait-for-it/master/wait-for-it.sh -O /usr/local/bin/wait-for-it \
   && chmod u+x /usr/local/bin/wait-for-it \
   && pip install --no-cache-dir -r requirements.txt
@@ -12,22 +12,12 @@ FROM hasura/graphql-engine:v1.0.0.cli-migrations AS hasura-migrations
 # /hasura-migrations is the default folder where the client will look for migrations to apply
 COPY ./backend/database-service/migrations /hasura-migrations
 
-FROM python:3.8-slim AS hasura-test
+FROM python:3.8-slim AS python-tests
 
 COPY --from=hasura-build /usr/local/bin/wait-for-it /usr/local/bin/wait-for-it
 COPY --from=hasura-build /usr/local/lib/python3.8/site-packages /usr/local/lib/python3.8/site-packages
 COPY --from=hasura-build /usr/local/bin /usr/local/bin
 COPY --from=hasura-migrations /bin/hasura-cli /usr/local/bin/hasura
-
-WORKDIR /app
-
-FROM hasura-test AS hasura-test-ci
-
-WORKDIR /app
-
-COPY ./backend/database-service .
-COPY ./backend/test-utils ./utils
-COPY ./shared/fixtures ./fixtures
 
 FROM python:3.8-slim AS integration-test-build
