--- backend/database-service/pytest.ini
@@ -1,5 +1,0 @@
-[pytest]
-addopts =
-    -p no:cacheprovider
-    --database-service-folder /app
-    --fixtures-folder /app/fixtures
