--- backend/database-service/tests/requirements.txt
@@ -1,5 +1,0 @@
-psycopg2-binary==2.8.4
-pytest==5.2.2
-sh==1.12.14
-SQLAlchemy==1.2.5
-SQLAlchemy-Utils==0.33.1
